Move auxflash handling to an extension trait
